Code_Correction_Part/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
requests
multidict==6.1.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
requests==2.32.3
typing-inspect==0.9.0
typing_extensions==4.12.2